requirements.txt
asgiref==3.2.7
astroid==2.4.1
certifi==2020.4.5.1
chardet==3.0.4
coinbase==2.1.0
colorama==0.4.3
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.6
django-allauth==0.41.0
django-filter==2.2.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
gunicorn==20.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
oauthlib==3.1.0
psycopg2==2.8.5
pycryptodome==3.9.7
pylint==2.5.2
python-dotenv==0.13.0
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
urllib3==1.25.9
whitenoise==5.0.1
wrapt==1.12.1
